refactor: remove fallback code for ECDH key derivation
