+++ jwa/ecdhes/ecdhes.go
@@ -4,6 +4,8 @@ package ecdhes
 
 import (
 	"crypto"
+	"crypto/ecdh"
+	"crypto/ecdsa"
 	"crypto/rand"
 	_ "crypto/sha256" // for crypto.SHA256
 	"fmt"
@@ -16,6 +18,8 @@ import (
 	"github.com/shogo82148/goat/jwk"
 	"github.com/shogo82148/goat/jwk/jwktypes"
 	"github.com/shogo82148/goat/keymanage"
+	"github.com/shogo82148/goat/x25519"
+	"github.com/shogo82148/goat/x448"
 )
 
 var alg = &Algorithm{
@@ -297,4 +301,51 @@ func (r *kdf) putUint32(v uint32) {
 	buf[2] = byte(v >> 8)
 	buf[3] = byte(v)
 	r.hash.Write(buf)
+}
+
+func deriveZ(priv, pub any) ([]byte, error) {
+	switch priv := priv.(type) {
+	case x25519.PrivateKey:
+		pubkey, ok := pub.(x25519.PublicKey)
+		if !ok {
+			return nil, fmt.Errorf("ecdhes: want x25519.PublicKey but got %T", pub)
+		}
+		privECDH, err := priv.ECDH()
+		if err != nil {
+			return nil, err
+		}
+		pubECDH, err := pubkey.ECDH()
+		if err != nil {
+			return nil, err
+		}
+		return privECDH.ECDH(pubECDH)
+	case x448.PrivateKey:
+		pubkey, ok := pub.(x448.PublicKey)
+		if !ok {
+			return nil, fmt.Errorf("ecdhes: want x447.PublicKey but got %T", pub)
+		}
+		return x448.X448(priv[:x448.SeedSize], pubkey)
+	case *ecdsa.PrivateKey:
+		pubkey, ok := pub.(*ecdsa.PublicKey)
+		if !ok {
+			return nil, fmt.Errorf("ecdhes: want *ecdsa.PrivateKey but got %T", pub)
+		}
+		privECDH, err := priv.ECDH()
+		if err != nil {
+			return nil, err
+		}
+		pubECDH, err := pubkey.ECDH()
+		if err != nil {
+			return nil, err
+		}
+		return privECDH.ECDH(pubECDH)
+	case *ecdh.PrivateKey:
+		pubkey, ok := pub.(*ecdh.PublicKey)
+		if !ok {
+			return nil, fmt.Errorf("ecdhes: want *ecdh.PublicKey but got %T", pub)
+		}
+		return priv.ECDH(pubkey)
+	default:
+		return nil, fmt.Errorf("ecdhes: unknown private key type: %T", priv)
+	}
 }
